# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tidarr API",
    "description": "Tidarr REST API for managing Tidal music downloads, queue management, synchronization, and Lidarr integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8484",
      "description": "Local Tidarr instance"
    }
  ],
  "security": [
    {
      "ApiKeyHeader": []
    },
    {
      "ApiKeyQuery": []
    },
    {
      "BearerToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "API key passed via X-Api-Key header (recommended for automation)"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "API key passed via query parameter"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/auth"
      }
    },
    "schemas": {
      "ContentType": {
        "type": "string",
        "enum": [
          "album",
          "track",
          "playlist",
          "artist",
          "video",
          "favorite_albums",
          "favorite_tracks",
          "favorite_playlists",
          "favorite_videos",
          "favorite_artists",
          "artist_videos",
          "mix"
        ],
        "description": "Type of Tidal content"
      },
      "QualityType": {
        "type": "string",
        "enum": [
          "low",
          "normal",
          "high",
          "max"
        ],
        "description": "Download quality. low=AAC-96, normal=AAC-320, high=16-bit FLAC, max=24-bit FLAC"
      },
...
      "SyncItem": {
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "type",
          "quality"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Tidal content ID or UUID"
          },
          "title": {
            "type": "string",
            "description": "Playlist or item title"
          },
          "url": {
            "type": "string",
            "description": "Full Tidal URL"
          },
          "type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "quality": {
            "$ref": "#/components/schemas/QualityType"
          },
          "artist": {
            "type": "string",
            "description": "Artist name (optional)"
          },
          "lastUpdate": {
            "type": "string",
            "description": "ISO 8601 timestamp of last sync"
          }
        }
      },
...
      "SaveSuccessResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "TiddlTomlResponse": {
        "type": "object",
        "required": [
          "toml"
        ],
        "properties": {
          "toml": {
            "type": "string",
            "description": "Raw TOML content of the tiddl configuration file"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/sync/save": {
      "post": {
        "operationId": "addToSyncList",
        "summary": "Add item to sync list",
        "description": "Add a playlist or other content to the synchronization list. Items in the sync list are automatically re-downloaded on a cron schedule (default: 3 AM daily).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/SyncItem"
                  }
                }
              },
              "examples": {
                "playlist": {
                  "summary": "Sync a playlist",
                  "value": {
                    "item": {
                      "id": "abc123-def456",
                      "title": "My Playlist",
                      "url": "https://listen.tidal.com/playlist/abc123-def456",
                      "type": "playlist",
                      "quality": "high"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item added to sync list"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/remove": {
      "delete": {
        "operationId": "removeFromSyncList",
        "summary": "Remove item from sync list",
        "description": "Remove a specific item from the synchronization list by its ID.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "description": "ID of the sync item to remove"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Item removed from sync list"
          },
          "400": {
            "description": "Invalid or missing ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tiddl/config": {
      "get": {
        "operationId": "getTiddlConfig",
        "summary": "Get Tiddl TOML configuration",
        "description": "Returns the raw TOML content of the tiddl configuration file (config.toml).",
        "responses": {
          "200": {
            "description": "Tiddl config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TiddlTomlResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "saveTiddlConfig",
        "summary": "Save Tiddl TOML configuration",
        "description": "Save the tiddl TOML configuration file content. This overwrites the entire config.toml file.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "toml"
                ],
                "properties": {
                  "toml": {
                    "type": "string",
                    "description": "Raw TOML content to save"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Config saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "TOML content must be a string"
          }
        }
      }
    }
  },
...
}